# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OrionOps API
  description: Enterprise Service Orchestration Platform - ITSM + ERP Extensions
  version: 1.0.0
  contact:
    name: OrionOps Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /problems:
    get:
      tags: [Problems]
      summary: List problems
      operationId: listProblems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list
    post:
      tags: [Problems]
      summary: Create problem
      operationId: createProblem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemRequest'
      responses:
        '201':
          description: Problem created
  /problems/{id}:
    get:
      tags: [Problems]
      summary: Get problem
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Problem details
    put:
      tags: [Problems]
      summary: Update problem
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Problem updated
    delete:
      tags: [Problems]
      summary: Delete problem
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted
  /problems/{id}/link-incident:
    patch:
      tags: [Problems]
      summary: Link incident to problem
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Incident linked
  /problems/{id}/root-cause:
    patch:
      tags: [Problems]
      summary: Set root cause
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Root cause set

  # ---- Changes ----
...
  /requests:
    get:
      tags: [Requests]
      summary: List service requests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list
    post:
      tags: [Requests]
      summary: Create service request
      responses:
        '201':
          description: Request created
  /requests/{id}:
    get:
      tags: [Requests]
      summary: Get service request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Request details
    put:
      tags: [Requests]
      summary: Update request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Requests]
      summary: Delete request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted
  /requests/{id}/approve:
    patch:
      tags: [Requests]
      summary: Approve request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Approved
  /requests/{id}/fulfill:
    patch:
      tags: [Requests]
      summary: Fulfill request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Fulfilled
  /requests/{id}/close:
    patch:
      tags: [Requests]
      summary: Close request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Closed

  # ---- CMDB ----
...
  /cmdb/services/{id}:
    get:
      tags: [CMDB]
      summary: Get service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Service details
    put:
      tags: [CMDB]
      summary: Update service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [CMDB]
      summary: Delete service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted

  # ---- SLA ----
...
  /knowledge:
    get:
      tags: [Knowledge]
      summary: List knowledge articles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list
    post:
      tags: [Knowledge]
      summary: Create article
      responses:
        '201':
          description: Article created
  /knowledge/{id}:
    get:
      tags: [Knowledge]
      summary: Get article
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Article details
    put:
      tags: [Knowledge]
      summary: Update article
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Knowledge]
      summary: Delete article
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted
  /knowledge/{id}/submit-for-review:
    patch:
      tags: [Knowledge]
      summary: Submit for review
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Submitted
  /knowledge/{id}/publish:
    patch:
      tags: [Knowledge]
      summary: Publish article
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Published

  # ---- Workflow ----
...
  /procurement/requests/{id}/submit:
    post:
      tags: [Procurement]
      summary: Submit purchase request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Submitted
  /procurement/requests/{id}/create-po:
    post:
      tags: [Procurement]
      summary: Create purchase order from request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '201':
          description: PO created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
      description: Page number (0-indexed)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
      description: Page size
...
    EntityIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Entity UUID
  schemas:
...
    ProblemRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        category:
          type: string
        serviceId:
          type: string
          format: uuid
        assigneeId:
          type: string
          format: uuid